refactor: 폴더구조변경, 코드정리
