+++ src/main/java/org/example/prumpt_be/dto/response/RateAvgDto.java
@@ -1,10 +1,7 @@
 package org.example.prumpt_be.dto.response;
 
-import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
 import lombok.Setter;
-import org.example.prumpt_be.domain.entity.PromptReviews;
 
 
 // 사용자 평균 평점 DTO
