+++ src/main/java/org/example/prumpt_be/repository/UsersRepository.java
@@ -1,8 +1,6 @@
 package org.example.prumpt_be.repository;
 
-import lombok.Getter;
-import lombok.Setter;
-import org.example.prumpt_be.domain.entity.Users;
+import org.example.prumpt_be.dto.entity.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
